Switch to a better method of preloading images
The previous method created Image objects, loaded the images on page
load, and then immediately released its reference to these images. What
I was seeing was that after a while the images themselves get GC'd
(particularly "man on goal" and "block on goal", which often aren't
displayed depending on board state), and we nevertheless have a lag as
they're loaded again.

The new method, suggested by [this helpful
site](http://perishablepress.com/3-ways-preload-images-css-javascript-ajax/),
sets each image to be the background of a div, but with the background
aligned so far off to the side it would never actually show up.

Another slightly cleaner method may be to just hold the Image objects
in an atom, but this seems to work.